Fixed label and format for transaction amount
