+++ test/fixtures/transaction_details.yml
@@ -639,7 +639,7 @@ cfd_annual:
   original_file_date: <%= 2.weeks.ago.to_date %>
   tcm_financial_year: "2021"
 
-pas:
+pas: &PAS_DEFAULTS
   transaction_header: pas
   sequence_number: 1
   customer_reference: "A16010889P"
@@ -700,6 +700,80 @@ pas:
   original_filename: 'PASYI00337'
   original_file_date: <%= 6.weeks.ago.to_date %>
   tcm_financial_year: "1718"
+  transaction_header: pas
+  sequence_number: 1
+  transaction_type: "I"
+  transaction_reference: <%= rand(36**8).to_s(36).upcase %>
+  related_reference: ""
+  currency_code: "GBP"
+  header_narrative: ""
+  header_attr_1: "20-AUG-2017"
+  header_attr_2: "Site"
+  header_attr_3: "Red St. Hill Farm"
+  header_attr_4: ""
+  header_attr_5: ""
+  header_attr_6: ""
+  header_attr_7: ""
+  header_attr_8: ""
+  header_attr_9: "Charge Period: "
+  header_attr_10: "21/04/2017 - 31/03/2018"
+  line_vat_code: ""
+  line_area_code: "Y05"
+  line_income_stream_code: "P"
+  line_context_code: "P1"
+  line_attr_1: ""
+  line_attr_2: ""
+  line_attr_3: ""
+  line_attr_4: ""
+  line_attr_5: ""
+  line_attr_6: ""
+  line_attr_7: ""
+  line_attr_8: ""
+  line_attr_9: ""
+  line_attr_10: ""
+  line_attr_11: "F 3.0"
+  line_attr_12: ""
+  line_attr_13: ""
+  line_attr_14: ""
+  line_attr_15: "Directly Associated Activity only"
+  line_quantity: 1
+  unit_of_measure: "Each"
+  generated_filename:
+  generated_file_at:
+  temporary_cessation: false
+  category:
+  charge_calculation:
+
+###########################################################
+# PAS - Retrospective transactions
+#
+
+pas_retro_invoice_1:
+  <<: *PAS_DEFAULTS
+  customer_reference: "A999222P"
+  transaction_date: <%= 2.months.ago %>
+  transaction_type: "I"
+  transaction_reference: "PAS0005811Y"
+  related_reference:
+  currency_code: "GBP"
+  header_attr_10: "21/04/2017 - 31/03/2018"
+  line_amount: 323260
+  line_description: "Subsistence Charge Permit No: VP3999DA"
+  unit_of_measure_price: 323260
+  status: "retrospective"
+  filename: "PASYI00337.dat"
+  reference_1: "VP3999DA"
+  reference_2: "ABC9988V"
+  reference_3: "AB1234ZZ"
+  temporary_cessation: false
+  category:
+  charge_calculation:
+  period_start: "21-Apr-2017"
+  period_end: "31 Mar 2018"
+  original_filename: 'PASYI00337'
+  original_file_date: <%= 6.weeks.ago.to_date %>
+  tcm_financial_year: "1718"
+
 
 pas_annual:
   transaction_header: pas
